+++ simulation/recruitmentprob.h
@@ -82,8 +82,8 @@ inline bool RecruitmentProb::intCompareGTEQ(const double param1, const double pa
         fracpart1 = modf(Store1, &intpart1);
         fracpart2 = modf(Store2, &intpart2);
 
-        intpart1Store = intpart1;
-        intpart2Store = intpart2;
+        intpart1Store = (int)intpart1;
+        intpart2Store = (int)intpart2;
 
         if (intpart1Store >= intpart2Store)
         {
